+++ Src-triggers/TrigTCP.cpp
@@ -17,7 +17,7 @@ static TrigTCP      *ME;
 
 void TrTCPWorker::run()
 {
-    const int   niq = viq.size();
+    const int   niq = (int)viq.size();
     bool        ok  = true;
 
     for(;;) {
@@ -53,7 +53,7 @@ bool TrTCPWorker::writeSome( int iq )
     if( !ME->nSampsFromCt( data, headCt, -LOOP_MS, S.js, S.ip ) )
         return false;
 
-    uint    size = data.size();
+    uint    size = (uint)data.size();
 
     if( !size )
         return true;
@@ -165,7 +165,7 @@ void TrigTCP::run()
     TrTCPShared                 shr( p );
     const int                   nPrbPerThd  = 2;
 
-    iqMax   = vS.size() - 1;
+    iqMax   = (int)vS.size() - 1;
     nThd    = 0;
 
     for( int iq0 = 0; iq0 < iqMax; iq0 += nPrbPerThd ) {
